Separate components into ./react-app/src/components and Storybook stories into ./react-app/src/stories
