+++ react-app/src/components/SearchBar.js
