+++ react-app/src/components/SearchButton.js
