+++ react-app/src/components/UserPanel.js
