+++ react-app/src/stories/Button.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Button } from "./Button";
+import { Button } from "../components/Button";
 
 export default {
   title: "Services Page Prototype/Button",
