+++ react-app/src/stories/Header.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Header from "./Header";
+import Header from "../components/Header";
 import * as UserPanelStories from "./UserPanel.stories";
 
 export default {
