+++ react-app/src/stories/Introduction.stories.mdx
@@ -1,12 +1,12 @@
 import { Meta } from '@storybook/addon-docs/blocks';
-import Code from './assets/code-brackets.svg';
-import Colors from './assets/colors.svg';
-import Comments from './assets/comments.svg';
-import Direction from './assets/direction.svg';
-import Flow from './assets/flow.svg';
-import Plugin from './assets/plugin.svg';
-import Repo from './assets/repo.svg';
-import StackAlt from './assets/stackalt.svg';
+import Code from '../components/assets/code-brackets.svg';
+import Colors from '../components/assets/colors.svg';
+import Comments from '../components/assets/comments.svg';
+import Direction from '../components/assets/direction.svg';
+import Flow from '../components/assets/flow.svg';
+import Plugin from '../components/assets/plugin.svg';
+import Repo from '../components/assets/repo.svg';
+import StackAlt from '../components/assets/stackalt.svg';
 
 <Meta title="Example/Introduction" />
 
@@ -73,7 +73,7 @@ import StackAlt from './assets/stackalt.svg';
     display: block;
     margin-bottom: 2px;
   }
-  
+
   .link-item img {
     height: 40px;
     width: 40px;
@@ -111,7 +111,7 @@ import StackAlt from './assets/stackalt.svg';
     display: inline-block;
   }
 
-  
+
 `}</style>
 
 # Welcome to Storybook
