+++ react-app/src/stories/Page.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Page } from "./Page";
+import { Page } from "../components/Page";
 import * as HeaderStories from "./Header.stories";
 
 export default {
