+++ react-app/src/stories/SearchBar.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { SearchBar } from "./SearchBar";
+import SearchBar from "../components/SearchBar";
 
 export default {
   title: "Services Page Prototype/SearchBar",
