+++ react-app/src/stories/Service.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Service } from "./Service";
+import { Service } from "../components/Service";
 
 export default {
   title: "Services Page Prototype/Service",
